requirements.txt
flask==3.1.0
flask-cors==5.0.1
yfinance==0.2.54
openai==1.68.0
pandas==2.2.3
ta==0.11.0
python-dotenv==1.1.0
cachetools==5.5.2
